feat/MSSDK-1948: Adjust Sentry integration for shared environments capability
